Update documentation/help.
When I extracted the Class superclass I renamed the _update_config()
method in controller.py to register() to make it callable from the
command line (as with the global classes). I've now updated the help()
method in the Controller class and the docs/intro.txt file to document
this feature.