Fix duplication bug in controller overrides.
If an override of a certain module's controllers was already
registered in a module's configuration file, that override
could accidentally be duplicated upon running

    magetool -os Mage_CertainModule create controller X

where X was a name which did not clash with that of any of
the module's existing controller's. I added a check for
previously defined overrides to fix this bug.