Make _create_class() work in any subdirectory.
I made a slight change to the _create_class() method in the Class
class so that it now takes advantage of the new path attribute
in the Module class. This allows you to make invocations such as

    magetool create model Order

even when you're one or more levels into a module directory (e.g.,
in the Block/ directory). Previously you could only make such
invocations when you were in the module directory itself.